renaming and navigationActions
