+++ ts/navigation/MessagesNavigator.ts
@@ -12,8 +12,8 @@ import {
 } from "../config";
 import EuCovidCertNavigator from "../features/euCovidCert/navigation/navigator";
 import EUCOVIDCERT_ROUTES from "../features/euCovidCert/navigation/routes";
-import MvlNavigator from "../features/mvl/navigatioon/navigator";
-import MVL_ROUTES from "../features/mvl/navigatioon/routes";
+import MvlNavigator from "../features/mvl/navigation/navigator";
+import MVL_ROUTES from "../features/mvl/navigation/routes";
 import MessageDetailScreen from "../screens/messages/MessageDetailScreen";
 import MessageRouterScreen from "../screens/messages/MessageRouterScreen";
 import MessagesHomeScreen from "../screens/messages/MessagesHomeScreen";
